backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet==3.0.3

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Data Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# File Handling
aiofiles==23.2.1
python-multipart==0.0.6
pillow==10.1.0
python-magic-bin==0.4.14

# Environment & Configuration
python-dotenv==1.0.0

# HTTP Client (for AI integration)
httpx==0.26.0

# Email Notifications
aiosmtplib==3.0.1
jinja2==3.1.2

# Development & Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# AI/ML - Updated to new Google Gen AI SDK
openai==1.7.0
google-genai>=0.1.0
fuzzywuzzy[speedup]==0.18.0

# Agent-specific dependencies
perplexity-api==1.0.5
selenium==4.15.0
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
pandas==2.1.4

# Utilities
python-slugify==8.0.1
pytz==2023.3

# AI Agent Dependencies - Updated
google-genai>=0.1.0
python-dotenv>=1.0.0
selenium>=4.15.0
requests>=2.31.0 